Set default network driver on all targets to ethernet
As this example by default is set to work with ethernet, the
default driver is set to ethernet for all targets. This is done to
allow using different network drivers only by editing mbed_app.json.

The config options for the pre Mbed OS-5.10 network connection method
have been removed as well.